requirements.txt
fastapi==0.124.4
uvicorn==0.38.0
SQLAlchemy==2.0.45
asyncpg==0.31.0
alembic==1.17.2
pydantic==2.12.5
pydantic-settings==2.12.0
python-jose==3.5.0
passlib[bcrypt]==1.7.4
fastapi-mail==1.6.1
httpx==0.28.1
openpyxl==3.1.5
python-multipart==0.0.20
email-validator==2.3.0
